feat(homepage): bring back the old version
