+++ packages/homepage/src/screens/home/hero/index.js
@@ -1,77 +1,11 @@
-import React, { useState } from 'react';
-import { motion, AnimatePresence } from 'framer-motion';
-import { useStaticQuery, graphql } from 'gatsby';
-import { track } from '@codesandbox/common/lib/utils/analytics';
+import React from 'react';
+import { motion } from 'framer-motion';
 
-import {
-  TemplateItem,
-  TemplateList,
-  HeroWrapper,
-  Title,
-  SubTitle,
-  ShowMoreIcon,
-  ShowMoreButton,
-  ShowMoreButtonLine,
-} from './elements';
-import Plus from '../../../assets/icons/Plus';
-
-const MAIN_TEMPLATES = [
-  'react-new',
-  'vanilla-vanilla',
-  'vue-vue',
-  'angular-angular',
-];
-
-const HIDE_LIST = ['p5js-9tgmp', 'codesandbox-appsapper-template-88wlq002r8'];
+import { HeroWrapper, Title, SubTitle } from './elements';
+import { CTATemplateAVersion } from './A-Version';
+import { ListTemplateBVersion } from './B-Version';
 
 export default () => {
-  const [showMore, setShowMore] = useState(false);
-
-  const {
-    allOfficialTemplate: { nodes },
-  } = useStaticQuery(graphql`
-    query OfficialTemplates {
-      allOfficialTemplate {
-        nodes {
-          id
-          author {
-            username
-          }
-          alias
-          custom_template {
-            color
-            id
-            icon_url
-          }
-          description
-          environment
-          title
-        }
-      }
-    }
-  `);
-
-  const mainTemplates = nodes
-    .filter(e => MAIN_TEMPLATES.includes(e.alias))
-    .sort((a, b) => {
-      if (MAIN_TEMPLATES.indexOf(a.alias) > MAIN_TEMPLATES.indexOf(b.alias)) {
-        return 1;
-      }
-
-      return -1;
-    });
-  const allNodesSort = nodes
-    .filter(e => !MAIN_TEMPLATES.includes(e.alias))
-    .filter(e => !HIDE_LIST.includes(e.alias))
-    .sort((a, b) => {
-      if (a.title > b.title) {
-        return 1;
-      }
-
-      return -1;
-    })
-    .slice(0, 28);
-
   return (
     <HeroWrapper>
       <motion.div
@@ -92,72 +26,8 @@ export default () => {
           web development.
         </SubTitle>
 
-        <TemplateList>
-          {mainTemplates.map((template, index) => (
-            <TemplateItem
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              transition={{ duration: (index + 1) * 0.3, delay: 1 }}
-              as={motion.li}
-              key={template.alias}
-              alias={template.alias}
-              title={template.title}
-              environment={template.environment}
-              iconUrl={template.custom_template.icon_url}
-            />
-          ))}
-        </TemplateList>
-
-        <AnimatePresence>
-          {showMore && (
-            <TemplateList
-              as={motion.div}
-              initial="collapsed"
-              animate="open"
-              exit="collapsed"
-              variants={{
-                open: { opacity: 1, height: 'auto' },
-                collapsed: { opacity: 0, height: 0 },
-              }}
-              transition={{ duration: 0.4, ease: 'easeInOut' }}
-            >
-              {allNodesSort.map((template, index) => (
-                <TemplateItem
-                  initial={{ opacity: 0 }}
-                  animate={{ opacity: 1 }}
-                  transition={{ duration: Math.min((index + 1) * 0.3, 1.6) }}
-                  as={motion.li}
-                  key={template.alias}
-                  alias={template.alias}
-                  title={template.title}
-                  environment={template.environment}
-                  iconUrl={template.custom_template.icon_url}
-                />
-              ))}
-            </TemplateList>
-          )}
-        </AnimatePresence>
-
-        <TemplateList>
-          <ShowMoreButtonLine
-            as={motion.div}
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            transition={{ duration: 0.3, delay: 1.4 }}
-          >
-            <ShowMoreButton
-              onClick={() => {
-                setShowMore(p => !p);
-                track('Homepage - Template show more', { open: showMore });
-              }}
-            >
-              <ShowMoreIcon active={showMore}>
-                <Plus />
-              </ShowMoreIcon>
-              {showMore ? 'Less Template' : 'More Templates'}
-            </ShowMoreButton>
-          </ShowMoreButtonLine>
-        </TemplateList>
+        <ListTemplateBVersion />
+        <CTATemplateAVersion />
       </motion.div>
     </HeroWrapper>
   );
